Extract PayloadSerializer interface
