+++ platforms/core-runtime/serialization/src/main/java/org/gradle/tooling/internal/provider/serialization/PayloadSerializer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,69 +16,15 @@
 
 package org.gradle.tooling.internal.provider.serialization;
 
-import org.apache.commons.io.IOUtils;
-import org.gradle.internal.Cast;
-import org.gradle.internal.UncheckedException;
-import org.gradle.internal.io.StreamByteBuffer;
 import org.gradle.internal.service.scopes.Scope;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 import javax.annotation.Nullable;
-import javax.annotation.concurrent.ThreadSafe;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 
-@ThreadSafe
 @ServiceScope(Scope.UserHome.class)
-public class PayloadSerializer {
-    private final PayloadClassLoaderRegistry classLoaderRegistry;
+public interface PayloadSerializer {
+    SerializedPayload serialize(@Nullable Object payload);
 
-    public PayloadSerializer(PayloadClassLoaderRegistry registry) {
-        classLoaderRegistry = registry;
-    }
-
-    public SerializedPayload serialize(@Nullable Object payload) {
-        if (payload == null) {
-            return new SerializedPayload(null, Collections.<byte[]>emptyList());
-        }
-
-        final SerializeMap map = classLoaderRegistry.newSerializeSession();
-        try {
-            StreamByteBuffer buffer = new StreamByteBuffer();
-            final ObjectOutputStream objectStream = new PayloadSerializerObjectOutputStream(buffer.getOutputStream(), map);
-
-            try {
-                objectStream.writeObject(payload);
-            } finally {
-                IOUtils.closeQuietly(objectStream);
-            }
-
-            Map<Short, ClassLoaderDetails> classLoaders = new HashMap<Short, ClassLoaderDetails>();
-            map.collectClassLoaderDefinitions(classLoaders);
-            return new SerializedPayload(classLoaders, buffer.readAsListOfByteArrays());
-        } catch (IOException e) {
-            throw UncheckedException.throwAsUncheckedException(e);
-        }
-    }
-
-    public @Nullable Object deserialize(SerializedPayload payload) {
-        if (payload.getSerializedModel().isEmpty()) {
-            return null;
-        }
-
-        final DeserializeMap map = classLoaderRegistry.newDeserializeSession();
-        try {
-            final Map<Short, ClassLoaderDetails> classLoaderDetails = Cast.uncheckedNonnullCast(payload.getHeader());
-            StreamByteBuffer buffer = StreamByteBuffer.of(payload.getSerializedModel());
-            final ObjectInputStream objectStream = new PayloadSerializerObjectInputStream(buffer.getInputStream(), getClass().getClassLoader(), classLoaderDetails, map);
-            return objectStream.readObject();
-        } catch (Exception e) {
-            throw UncheckedException.throwAsUncheckedException(e);
-        }
-    }
-
+    @Nullable
+    Object deserialize(SerializedPayload payload);
 }
